Update Angular dependencies to latest versions
Update all Angular related dependencies along with the script builds in the "tic-tac-toe" project. The upgraded versions ensure that the project benefits from the latest improvements and bug fixes in the Angular framework and associated libraries. Additionally, the "@ampproject/remapping" module has been upgraded to 2.2.1 for maintaining the integrity and for the better development experience.