Upgrade Angular dependencies to v17.1.0 in tic-tac-toe project
The commit includes updating the various Angular related dependencies to v17.1.0 in the "tic-tac-toe" project. The diff reflects upgrades to both dependencies and devDependencies specified in package-lock.json. This helps to keep project up-to-date with the latest advancements and bug fixes offered by Angular. Plus, the "@ampproject/remapping" module is updated ensuring better development experience.